+++ src/HtmlFragment.php
@@ -12,7 +12,7 @@ class HtmlFragment extends HtmlDocument
      */
     protected $handleEncoding = false;
 
-    function loadEmpty(array $properties = null): void
+    function loadEmpty(?array $properties = null): void
     {
         $this->loadString('', null, $properties);
     }
@@ -38,7 +38,7 @@ HTML
         );
     }
 
-    function save(\DOMNode $contextNode = null, bool $childrenOnly = false): string
+    function save(?\DOMNode $contextNode = null, bool $childrenOnly = false): string
     {
         if ($contextNode === null) {
             $contextNode = $this->getBody();
@@ -48,7 +48,7 @@ HTML
         return parent::save($contextNode, $childrenOnly);
     }
 
-    function query(string $expression, \DOMNode $contextNode = null, bool $registerNodeNs = true): \DOMNodeList
+    function query(string $expression, ?\DOMNode $contextNode = null, bool $registerNodeNs = true): \DOMNodeList
     {
         // if no context node has been given, assume <body>
         if ($contextNode === null) {